+++ R/apply_niche.R
@@ -8,7 +8,7 @@ apply_niche = function(x, niche_def, gc){
   #' @param gc function, stands for "gradient change". Specifies how the gradient changes, e.g. with time. Must be vectorized, meaning if given a vector, it must return a vector of equal length.
   #'
   #' @description
-    #' Models niches by removing events (fossil occurrences) or specimens when they are outside of their niche. For event type data, this is done using the function `thin`, for `pre_paleoTS` this is done by applying the function `prob_remove` on the specimens.
+    #' Models niches by removing events (fossil occurrences) or specimens when they are outside of their niche. For event type data, this is done using the function `thin`, for `pre_paleoTS` this is done by applying the function `prob_remove` on the specimens. For `fossils` objects produced by the `FossilSim` package, fossil observations are removed according to the specified niche
     #' Combines the functions `niche_def` and `gc` ("gradient change") to determine how the taxons' collection probability changes with time/position. This is done by composing `niche_def` and `gc`. The result is then used to remove events/specimens in `x`.
   #'
   #' @returns for a numeric vector input, returns a numeric vector, timing/location of events (e.g. fossil ages/locations) preserved after the niche model is applied. For a `pre_paleoTS` object as input, returns a `pre_paleoTS` object with specimens removed according to the niche model. For a `fossils` object, returns a `fossils` object with some occurrences removed according to the niche definition
@@ -59,6 +59,19 @@ apply_niche = function(x, niche_def, gc){
     #'  # bc the taxon does not appear everywhere
     #'  # and there are fewer specimens per sampling site
     #'
+    #'  ### example for fossils objects
+    #'  # we reuse the niche definition and gradient change from above
+    #'  # simulate tree
+    #'  t = ape::rlineage(birth = 2, death = 0, Tmax = 2)
+    #'  # create fossils object
+    #'  f = FossilSim::sim.fossils.poisson(rate = 2, tree = t)
+    #'  # plot fossils along tree before niche model is applied
+    #'  FossilSim:::plot.fossils(f, tree = t)
+    #'  # introduce niche model
+    #'  f_mod = apply_niche(f, niche_def, gc)
+    #'  # plot fossils along tree after introduction of niche model
+    #'  FossilSim:::plot.fossils(f_mod, tree = t)
+    #'  # note how only fossils in the interval where environmental conditions are suitable are preserved
 
  UseMethod("apply_niche")
 
@@ -92,7 +105,7 @@ apply_niche.fossils = function(x, niche_def, gc){
   #'
   change_in_niche = function(y) niche_def(gc(y))
 
-  x_val = 0.5(x$hmin + x$hmax)
+  x_val = 0.5*(x$hmin + x$hmax)
   if (any(x$hmax != x$hmin)){
     warning("Fossils are asociated with age uncertainty. Using midpoint of min and max ages to determine removal probability")
   }