+++ tests/testthat/test_apply_niche.R
@@ -21,6 +21,11 @@ test_that("for pre_paleoTS, returns identical result for trivial niche", {
 })
 
 test_that("for fossils objects, returns indentical result for trivial niche",{
-  # for trivial niche is the identity function
+  niche_def = function(x) rep(1, length(x))
+  gc = function(x) rep(1, length(x))
+  t = ape::rlineage(birth = 1, death = 0, Tmax = 1)
+  rate = 2
+  f = FossilSim::sim.fossils.poisson(rate = rate, tree = t)
+  expect_equal(apply_niche(f, niche_def, gc), f)
 })